cast num_chars to unsigned size_t

+++ src/libmongoc/src/mongoc/mongoc-scram.c
@@ -1039,7 +1039,7 @@ _mongoc_sasl_prep_impl (const char *name, const char *in_utf8, bson_error_t *err
 
    /* convert to unicode. */
    BSON_ASSERT (bson_cmp_less_equal_su (num_chars, SIZE_MAX / sizeof (uint32_t) - 1u));
-   utf8_codepoints = bson_malloc (sizeof (uint32_t) * (num_chars + 1)); /* add one for trailing 0 value. */
+   utf8_codepoints = bson_malloc (sizeof (uint32_t) * ((size_t) num_chars + 1u)); /* add one for trailing 0 value. */
    const char *c = in_utf8;
 
    for (size_t i = 0; i < num_chars; ++i) {
